allow disabling

+++ packages/playwright/src/plugins/webServerPlugin.ts
@@ -108,6 +108,9 @@ export class WebServerPlugin implements TestRunnerPlugin {
         if (process.platform === 'win32')
           throw new Error('Graceful shutdown is not supported on Windows');
 
+        if (this._options.shutdownTimeout === 0)
+          throw new Error('skip graceful shutdown');
+
         const success = launchedProcess.kill('SIGINT');
         if (!success)
           throw new Error(`SIGINT didn't succeed, fall back to non-graceful shutdown`);
